Simplify Scala 3 macro implementation
