+++ scalatest/src/test/scala/user/org/mockito/stubbing/DefaultAnswerTest.scala
@@ -6,7 +6,7 @@ import org.mockito.stubbing.DefaultAnswer
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
-import org.scalatest.{OptionValues, TryValues}
+import org.scalatest.{ OptionValues, TryValues }
 import user.org.mockito.stubbing.DefaultAnswerTest.*
 
 object DefaultAnswerTest {
